git-svn-id: https://svn.apache.org/repos/asf/hive/branches/tez@1536823 13f79535-47bb-0310-9956-ffa450edef68

+++ testutils/ptest2/src/test/java/org/apache/hive/ptest/execution/TestReportParser.java
@@ -50,10 +50,12 @@ public class TestReportParser {
   public void test() throws Exception {
     File reportDir = new File("src/test/resources/test-outputs");
     for(File file : reportDir.listFiles()) {
-      if(file.getName().endsWith(".xml")) {
-        Files.copy(file, new File(baseDir, "TEST-" + file.getName()));
-      } else {
-        Files.copy(file, new File(baseDir, file.getName()));
+      if(file.isFile()) {
+        if(file.getName().endsWith(".xml")) {
+          Files.copy(file, new File(baseDir, "TEST-" + file.getName()));
+        } else {
+          Files.copy(file, new File(baseDir, file.getName()));
+        }
       }
     }
     JUnitReportParser parser = new JUnitReportParser(LOG, baseDir);
